Task 3/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
chartjs==1.2